Refactor Pandas Axis Partition objects to avoid duplicate code
* Create PandasOnXAxisPartition abstract class to contain common code
* Remove common code between `PandasOnRay` and `PandasOnPython` `AxisPartitions`
* Register a custom serializer for Ray to fit with the new code
