+++ modin/engines/ray/pandas_on_ray/axis_partition.py
@@ -2,15 +2,13 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import pandas
 import ray
 
-from modin.engines.base.axis_partition import BaseAxisPartition
-from modin.data_management.utils import split_result_of_axis_func_pandas
+from modin.engines.base.axis_partition import PandasOnXAxisPartition
 from .remote_partition import PandasOnRayRemotePartition
 
 
-class PandasOnRayAxisPartition(BaseAxisPartition):
+class PandasOnRayAxisPartition(PandasOnXAxisPartition):
     def __init__(self, list_of_blocks):
         # Unwrap from BaseRemotePartition object for ease of use
         self.list_of_blocks = [obj.oid for obj in list_of_blocks]
@@ -18,71 +16,38 @@ class PandasOnRayAxisPartition(BaseAxisPartition):
     partition_type = PandasOnRayRemotePartition
     instance_type = ray.ObjectID
 
-    def apply(
-        self,
-        func,
-        num_splits=None,
-        other_axis_partition=None,
-        maintain_partitioning=True,
-        **kwargs
+    @classmethod
+    def deploy_axis_func(
+        cls, axis, func, num_splits, kwargs, maintain_partitioning, *partitions
     ):
-        """Applies func to the object in the plasma store.
-
-        See notes in Parent class about this method.
-
-        Args:
-            func: The function to apply.
-            num_splits: The number of times to split the result object.
-            other_axis_partition: Another `PandasOnRayAxisPartition` object to apply to
-                func with this one.
-            maintain_partitioning: Whether or not to keep the partitioning in the same
-                orientation as it was previously. This is important because we may be
-                operating on an individual AxisPartition and not touching the rest.
-                In this case, we have to return the partitioning to its previous
-                orientation (the lengths will remain the same). This is ignored between
-                two axis partitions.
-
-        Returns:
-            A list of `RayRemotePartition` objects.
-        """
-        if num_splits is None:
-            num_splits = len(self.list_of_blocks)
-
-        if other_axis_partition is not None:
-            return self._wrap_partitions(
-                deploy_ray_func_between_two_axis_partitions._remote(
-                    args=(self.axis, func, num_splits, len(self.list_of_blocks), kwargs)
-                    + tuple(self.list_of_blocks + other_axis_partition.list_of_blocks),
-                    num_return_vals=num_splits,
-                )
-            )
-
-        args = [self.axis, func, num_splits, kwargs, maintain_partitioning]
-        args.extend(self.list_of_blocks)
-        return self._wrap_partitions(
-            deploy_ray_axis_func._remote(args, num_return_vals=num_splits)
+        return deploy_ray_func._remote(
+            args=(
+                PandasOnXAxisPartition.deploy_axis_func,
+                axis,
+                func,
+                num_splits,
+                kwargs,
+                maintain_partitioning,
+                *partitions,
+            ),
+            num_return_vals=num_splits,
         )
 
-    def shuffle(self, func, lengths, **kwargs):
-        """Shuffle the order of the data in this axis based on the `lengths`.
-
-        Extends `BaseAxisPartition.shuffle`.
-
-        Args:
-            func: The function to apply before splitting.
-            lengths: The list of partition lengths to split the result into.
-
-        Returns:
-            A list of RemotePartition objects split by `lengths`.
-        """
-        num_splits = len(lengths)
-        # We add these to kwargs and will pop them off before performing the operation.
-        kwargs["manual_partition"] = True
-        kwargs["_lengths"] = lengths
-        args = [self.axis, func, num_splits, kwargs, False]
-        args.extend(self.list_of_blocks)
-        return self._wrap_partitions(
-            deploy_ray_axis_func._remote(args, num_return_vals=num_splits)
+    @classmethod
+    def deploy_func_between_two_axis_partitions(
+            cls, axis, func, num_splits, len_of_left, kwargs, *partitions
+    ):
+        return deploy_ray_func._remote(
+            args=(
+                PandasOnXAxisPartition.deploy_func_between_two_axis_partitions,
+                axis,
+                func,
+                num_splits,
+                len_of_left,
+                kwargs,
+                *partitions
+            ),
+            num_return_vals=num_splits,
         )
 
 
@@ -105,98 +70,5 @@ class PandasOnRayRowPartition(PandasOnRayAxisPartition):
 
 
 @ray.remote
-def deploy_ray_axis_func(
-    axis, func, num_splits, kwargs, maintain_partitioning, *partitions
-):
-    """Deploy a function along a full axis in Ray.
-
-    Args:
-        axis: The axis to perform the function along.
-        func: The function to perform.
-        num_splits: The number of splits to return
-            (see `split_result_of_axis_func_pandas`)
-        kwargs: A dictionary of keyword arguments.
-        maintain_partitioning: If True, keep the old partitioning if possible.
-            If False, create a new partition layout.
-        partitions: All partitions that make up the full axis (row or column)
-
-    Returns:
-        A list of Pandas DataFrames.
-    """
-    # Pop these off first because they aren't expected by the function.
-    manual_partition = kwargs.pop("manual_partition", False)
-    lengths = kwargs.pop("_lengths", None)
-
-    dataframe = pandas.concat(partitions, axis=axis, copy=False)
-    result = func(dataframe, **kwargs)
-    if isinstance(result, pandas.Series):
-        if num_splits == 1:
-            return result
-        return [result] + [pandas.Series([]) for _ in range(num_splits - 1)]
-
-    if manual_partition:
-        # The split function is expecting a list
-        lengths = list(lengths)
-    # We set lengths to None so we don't use the old lengths for the resulting partition
-    # layout. This is done if the number of splits is changing or we are told not to
-    # keep the old partitioning.
-    elif num_splits != len(partitions) or not maintain_partitioning:
-        lengths = None
-    else:
-        if axis == 0:
-            lengths = [len(part) for part in partitions]
-            if sum(lengths) != len(result):
-                lengths = None
-        else:
-            lengths = [len(part.columns) for part in partitions]
-            if sum(lengths) != len(result.columns):
-                lengths = None
-    return split_result_of_axis_func_pandas(axis, num_splits, result, lengths)
-
-
-@ray.remote
-def deploy_ray_func_between_two_axis_partitions(
-    axis, func, num_splits, len_of_left, kwargs, *partitions
-):
-    """Deploy a function along a full axis between two data sets in Ray.
-
-    Args:
-        axis: The axis to perform the function along.
-        func: The function to perform.
-        num_splits: The number of splits to return
-            (see `split_result_of_axis_func_pandas`).
-        len_of_left: The number of values in `partitions` that belong to the
-            left data set.
-        kwargs: A dictionary of keyword arguments.
-        partitions: All partitions that make up the full axis (row or column)
-            for both data sets.
-
-    Returns:
-        A list of Pandas DataFrames.
-    """
-    lt_frame = pandas.concat(list(partitions[:len_of_left]), axis=axis, copy=False)
-    rt_frame = pandas.concat(list(partitions[len_of_left:]), axis=axis, copy=False)
-
-    result = func(lt_frame, rt_frame, **kwargs)
-    return split_result_of_axis_func_pandas(axis, num_splits, result)
-
-
-@ray.remote
-def deploy_ray_shuffle_func(axis, func, numsplits, kwargs, *partitions):
-    """Deploy a function that defines the partitions along this axis.
-
-    Args:
-        axis:
-        func:
-        numsplits:
-        kwargs:
-        partitions:
-
-    Returns:
-        A list of Pandas DataFrames.
-    """
-    dataframe = pandas.concat(partitions, axis=axis, copy=False)
-    result = func(dataframe, numsplits=numsplits, **kwargs)
-
-    assert isinstance(result, list)
-    return result
+def deploy_ray_func(func, *args):
+    return func(*args)
